[feat]: FlatScene フォールバック UI とパネル共通化
Liquid 非対応環境向け Flat レイアウトと Convert/Result/CenterPeek の共有コンテンツを追加する。
