+++ src/features/converter/ui/ResultPanel.tsx
@@ -4,7 +4,6 @@ import {
   GlassContainer,
   Html,
 } from '@liquid-dom/react'
-import { formatFileSize } from '../../../lib/formatFileSize'
 import type { FfmpegConvertState } from '../hooks/useFfmpegConvert'
 import type { VideoBackdropState } from '../hooks/useVideoBackdrop'
 import {
@@ -14,6 +13,7 @@ import {
   RESULT_GLASS_CONTAINER_PROPS,
   type ResultPanelState,
 } from '../types'
+import { ResultPanelContent } from './ResultPanelContent'
 import './converter.css'
 
 type ResultPanelProps = {
@@ -28,83 +28,24 @@ const resolvePanelState = (convertState: FfmpegConvertState): ResultPanelState =
   return 'idle'
 }
 
-/** 右（または下）の結果表示パネル */
+/** 右（または下）の結果表示パネル（Liquid DOM） */
 export const ResultPanel = ({
   backdrop,
   convertState,
   panelWidth = PANEL_WIDTH,
 }: ResultPanelProps) => {
   const panelState = resolvePanelState(convertState)
-  const gifPercentage =
-    backdrop.videoSize && convertState.result
-      ? ((convertState.result.gifSize / backdrop.videoSize) * 100).toFixed(1)
-      : null
-  const isSizeUp =
-    backdrop.videoSize && convertState.result
-      ? convertState.result.gifSize > backdrop.videoSize
-      : false
 
   return (
     <Frame width={panelWidth} minHeight={PANEL_MIN_HEIGHT}>
       <GlassContainer {...RESULT_GLASS_CONTAINER_PROPS}>
         <Glass {...GLASS_SHAPE_PROPS}>
           <Html sizing="fill">
-            <div
-              className={`panel-html panel-html--result panel-html--${panelState}`}
-            >
-              <h2>変換結果</h2>
-
-              {panelState === 'idle' && (
-                <div className="result-idle">
-                  <span className="result-idle__icon" aria-hidden>
-                    ✦
-                  </span>
-                  <p className="placeholder">
-                    変換が完了するとここに結果が表示されます
-                  </p>
-                </div>
-              )}
-
-              {panelState === 'error' && (
-                <div className="result-error">
-                  <p className="error-message">{convertState.error}</p>
-                </div>
-              )}
-
-              {panelState === 'success' && convertState.result && (
-                <div className="result-success">
-                  {backdrop.videoSize !== null && (
-                    <p className="stats">
-                      <strong>元の動画</strong>{' '}
-                      {formatFileSize(backdrop.videoSize)}
-                      <br />
-                      <strong>GIF</strong>{' '}
-                      {formatFileSize(convertState.result.gifSize)}
-                      {gifPercentage && (
-                        <>
-                          {' '}
-                          <span className={isSizeUp ? 'up' : 'down'}>
-                            ({isSizeUp ? '▲' : '▼'} {gifPercentage}%)
-                          </span>
-                        </>
-                      )}
-                    </p>
-                  )}
-
-                  <div className="preview">
-                    <img src={convertState.result.gifUrl} alt="変換後の GIF プレビュー" />
-                  </div>
-
-                  <a
-                    className="download-link"
-                    href={convertState.result.gifUrl}
-                    download="output.gif"
-                  >
-                    GIF をダウンロード
-                  </a>
-                </div>
-              )}
-            </div>
+            <ResultPanelContent
+              backdrop={backdrop}
+              convertState={convertState}
+              panelState={panelState}
+            />
           </Html>
         </Glass>
       </GlassContainer>
